+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Removed
+
+- Don't delete branches twice. Remove pubspec_override.yaml before publishing
+
 ## 11.2.2 - 2026-07-29
 
 ### Changed